+++ TubesPBO/src/model/Person.java
@@ -1,14 +1,14 @@
 package model;
 
 public class Person {
-    private long id;
+    private int id;
     private String password;
     private String name;
     private String email;
     private String phone;
     private String picPath;
 
-    public Person(long id, String password, String name, String email, String phone, String picPath) {
+    public Person(int id, String password, String name, String email, String phone, String picPath) {
         this.id = id;
         this.password = password;
         this.name = name;
@@ -17,11 +17,11 @@ public class Person {
         this.picPath = picPath; 
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
